requirements/base.txt
Django==1.7
Fabric==1.10.0
factory-boy==2.4.1
Pillow==2.5.3
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argh==0.25.0
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.15
croniter==0.3.5
django-authtools==1.0.0
django-bootstrap-form==3.1
django-braces==1.4.0
django-classy-tags==0.5.1
django-cors-headers==1.0.0
django-crispy-forms==1.4.0
django-extensions==1.4.0
django-grappelli==2.5.4
django-sekizai==0.7
django-smart-selects==1.0.9
django-stronghold==0.2.6
django-tables2==0.15.0
djangorestframework==3.0.3
ecdsa==0.11
fabtools==0.19.0
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
GitPython==0.3.0-beta2
graphviz==0.4.1
greenlet==0.4.4
jenkinsapi==0.2.25
kombu==3.0.23
logan==0.5.10
paramiko==1.15.1
pathtools==0.1.2
pycrypto==2.6.1
pydot==1.0.2
pygraphviz==1.2
pyparsing==1.5.7
python-dateutil==2.2
pytz==2014.7
requests==2.4.1
simplejson==3.6.3
six==1.8.0
sqlparse==0.1.12
virtualenv==1.11.6
watchdog==0.8.1
wsgiref==0.1.2
svn==0.3.25
django-celery
beautifulsoup4
python-jenkins
